Added function to remove system dependencies

+++ .assertions/dependency_manager/modules/systemlib/install.sh
@@ -39,6 +39,11 @@ fi
 
 echo "Info: found file '$LIBRARY_FILE'" 1>&2
 
+if [ -f "$DEPENDENCIES_LIB_DIR/$FULL_LIBRARY_NAME" ]; then
+	echo "Error: library already added to dependencies"
+	exit 1
+fi
+
 echo "Info: linking dependency to '$DEPENDENCIES_LIB_DIR'" 1>&2
 ln -s "$LIBRARY_FILE" "$DEPENDENCIES_LIB_DIR/$FULL_LIBRARY_NAME"
